Changed NamedTuples to classes with constructors for automatic error checking and simplified extensibility.
The idea is to allow for the inclusion of convenience constructors later on that won't have to be modified at all even if the definitions of the classes are later modified. Having the class's main constructor perform error checking allows for many methods to use those capabilities without having to reinvent the wheel. This is useful if we want to later construct objects from non-dictionaries.